+++ gluecodium/src/test/resources/smoke/listeners/output/android-kotlin/com/example/smoke/InternalListener.kt
@@ -8,7 +8,7 @@
 package com.example.smoke
 
 
-internal interface InternalListener {
+internal fun interface InternalListener {