+++ gluecodium/src/test/resources/smoke/nesting/output/android-kotlin/com/example/smoke/OuterClassWithInheritance.kt
@@ -38,7 +38,7 @@ class OuterClassWithInheritance : ParentClass {
         }
     }
 
-    interface InnerInterface {
+    fun interface InnerInterface {