+++ gluecodium/src/test/resources/smoke/skip/output/android-kotlin/com/example/smoke/SkipTagsInKotlin.kt
@@ -8,7 +8,9 @@
 package com.example.smoke
 
 
-interface SkipTagsInKotlin {
+fun interface SkipTagsInKotlin {
+
+
 
     fun dontSkipTagged() : Unit